file listing improvements and move launcher props
